% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{NaiveBayes}
\alias{NaiveBayes}
\title{Categorical Naive Bayes Classifier}
\description{
This class implements the Categorical Naive Bayes algorithm for categorical data.
}
\section{Usage}{

\preformatted{nb <- NaiveBayes$new(data, features, target)
nb$fit(X, y)
predictions <- nb$predict(new_data)}


\preformatted{nb <- NaiveBayes$new(data, features, target)
nb$fit(X, y)}


\preformatted{predictions <- nb$predict(new_data)}


\preformatted{probs <- nb$predict_proba(new_data)}


\preformatted{nb$print()}


\preformatted{nb$summary()}
}

\section{Methods}{

\preformatted{initialize(data, features, target)
fit(X, y, epsilon = 0.001)
predict(X)
predict_proba(X)
print()
summary()
check.numerical()
numerical.to.categorical(column, column_number)
transform.categorical(X)}
}

\section{Author}{

Arthur Parmentier
Fatou NDIENG
}

\examples{
\dontrun{
# Example usage
nb <- NaiveBayes$new(data, features, target)
nb$fit(X, y)
predictions <- nb$predict(new_data)
}


## ------------------------------------------------
## Method `NaiveBayes$fit`
## ------------------------------------------------

\dontrun{
# Example usage
nb <- NaiveBayes$new(data, features, target)
nb$fit(X, y)
}


## ------------------------------------------------
## Method `NaiveBayes$predict`
## ------------------------------------------------

\dontrun{
# Example usage
predictions <- nb$predict(new_data)
}


## ------------------------------------------------
## Method `NaiveBayes$predict_proba`
## ------------------------------------------------

\dontrun{
# Example usage
probs <- nb$predict_proba(new_data)
}


## ------------------------------------------------
## Method `NaiveBayes$print`
## ------------------------------------------------

\dontrun{
# Example usage
nb$print()
}


## ------------------------------------------------
## Method `NaiveBayes$summary`
## ------------------------------------------------

\dontrun{
# Example usage
nb$summary()
}

}
\seealso{
\url{https://your_documentation_link}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The training dataset.}

\item{\code{features}}{The explanatory variables.}

\item{\code{target}}{The target variable.}

\item{\code{prior.probabilities}}{Prior probabilities of classes.}

\item{\code{conditional.probabilities}}{Conditional probabilities of features.}

\item{\code{X}}{Training data.}

\item{\code{y}}{Target variable in the training data.}

\item{\code{classes}}{Unique classes of the target variable.}

\item{\code{groupes}}{Groups used for transforming numerical variables into categorical.}

\item{\code{nb.is.numeric}}{Indicator for the presence of numerical columns.}

\item{\code{epsilon}}{Epsilon parameter used in calculations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NaiveBayes-new}{\code{NaiveBayes$new()}}
\item \href{#method-NaiveBayes-fit}{\code{NaiveBayes$fit()}}
\item \href{#method-NaiveBayes-predict}{\code{NaiveBayes$predict()}}
\item \href{#method-NaiveBayes-predict_proba}{\code{NaiveBayes$predict_proba()}}
\item \href{#method-NaiveBayes-print}{\code{NaiveBayes$print()}}
\item \href{#method-NaiveBayes-summary}{\code{NaiveBayes$summary()}}
\item \href{#method-NaiveBayes-check.numerical}{\code{NaiveBayes$check.numerical()}}
\item \href{#method-NaiveBayes-numerical.to.categorical}{\code{NaiveBayes$numerical.to.categorical()}}
\item \href{#method-NaiveBayes-transform.categorical}{\code{NaiveBayes$transform.categorical()}}
\item \href{#method-NaiveBayes-clone}{\code{NaiveBayes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-new"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$new(data, features, target)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$fit(X, y, epsilon = 0.001)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The training data.}

\item{\code{y}}{The target variable.}

\item{\code{epsilon}}{Epsilon parameter for smoothing (default is 0.001).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Example usage
nb <- NaiveBayes$new(data, features, target)
nb$fit(X, y)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-predict"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{New data for prediction.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with predictions added as a new column.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Example usage
predictions <- nb$predict(new_data)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$predict_proba(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{New data for which to calculate class membership probabilities.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with class membership probabilities for each observation.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Example usage
probs <- nb$predict_proba(new_data)
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-print"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints a summary to the console.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Example usage
nb$print()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-summary"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints a summary to the console.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
# Example usage
nb$summary()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-check.numerical"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-check.numerical}{}}}
\subsection{Method \code{check.numerical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$check.numerical()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-numerical.to.categorical"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-numerical.to.categorical}{}}}
\subsection{Method \code{numerical.to.categorical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$numerical.to.categorical(column, column_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-transform.categorical"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-transform.categorical}{}}}
\subsection{Method \code{transform.categorical()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$transform.categorical(X)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
